Move all the blocks components to a Common CaseStudyBlocks component
